Implemented Sign in With Google and Sign In With Email
Set up Firebase applications and Google Cloud OAth clients for both iOS and Android. Added GoogleService-Info.plist, which handles iOS, and google-services.json for Android. Package.json files where modified as a result of necessary library installations. Made some tweaks in App.tsx to initliaze the authentication through google and made it so that by default it currently takes us to the Login screen. This should be updated when we create the Sign Up screen and actual app navigation. In firebase.js added the Client ID for both clients as comments.